backend/property_service/requirements.txt
Django>=4.2
psycopg[binary]==3.2.2
gunicorn
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-environ
django-celery-beat
django-celery-results
django-redis
pika         
redis         
celery   
Pillow 